Better memory management in ProcessBlock()
In the shunting area on line 2031, where blocks that do not have their previous block yet are handled, there is a special case for proof-of-stake blocks. That case can fail and a return false is called. Above that there is a creation of a CBlock by new. If the return false is reached, that CBlock* that was allocated space is not handled. All I added was a delete pblock2; to handle memory better.